[Update] 삭제 여부 한 번 더 묻기(education & project mvp)

See merge request ai_track/class_04/web_project/team6/team6!51

+++ front/src/components/project/ProjectCard.js
@@ -4,10 +4,12 @@ import MoreHorizIcon from "@mui/icons-material/MoreHoriz";
 import EditIcon from "@mui/icons-material/Edit";
 import DeleteIcon from "@mui/icons-material/Delete";
 import * as Api from "../../api";
+import AlertDialog from "../utils/AlertDialog";
 
 function ProjectCard({ project, setClickEditBtn, isEditable, setProjects }) {
   const [anchorEl, setAnchorEl] = useState(null); // Menu Element를 가리킵니다.
   const [isOpen, setIsOpen] = useState(null); // Menu Element의 Open 상태를 저장합니다.
+  const [isDeleting, setIsDeleting] = useState(false); // 삭제 여부를 저장합니다.
 
   const handleClick = (event) => {
     setAnchorEl(event.currentTarget);
@@ -18,13 +20,15 @@ function ProjectCard({ project, setClickEditBtn, isEditable, setProjects }) {
     setIsOpen(false);
   };
 
-  const DelBtnClickHandler = async () => {
-    // projects로 DELETE 요청을 보내 프로젝트를 삭제합니다.
-    await Api.delete("projects", project.id);
+  const DelBtnClickHandler = async (isDeleting) => {
+    if (isDeleting) {
+      // projects로 DELETE 요청을 보내 프로젝트를 삭제합니다.
+      await Api.delete("projects", project.id);
 
-    // project-lists/유저id로 GET 요청을 보내 업데이트 사항이 반영된 프로젝트를 새로 저장합니다.
-    const res = await Api.get("project-lists", project.user.id);
-    setProjects(res.data);
+      // project-lists/유저id로 GET 요청을 보내 업데이트 사항이 반영된 프로젝트를 새로 저장합니다.
+      const res = await Api.get("project-lists", project.user.id);
+      setProjects(res.data);
+    }
   };
 
   return (
@@ -56,12 +60,15 @@ function ProjectCard({ project, setClickEditBtn, isEditable, setProjects }) {
                   <Button
                     color="error"
                     startIcon={<DeleteIcon />}
-                    onClick={DelBtnClickHandler}
+                    onClick={() => setIsDeleting(true)}
                   >
                     삭제
                   </Button>
                 </MenuItem>
               </Menu>
+            )}
+            {isDeleting && (
+              <AlertDialog checkDeleteComplete={DelBtnClickHandler} />
             )}
           </>
         )}
